- Flow working! - Now TOFIX raspberry, not taking 2 sinks (clones the first sink in the metadata)

+++ server/Home.Server.App/Sinks/FlowSInk.cs
@@ -39,7 +39,7 @@ namespace Lucky.Home.Sinks
             /// <summary>
             /// Count in the time unit (count/seconds)
             /// </summary>
-            public UInt32 Frequency;
+            public UInt16 Frequency;
         }
 
         /// <summary>
